+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
  * Add an `env` function to the expression language provider
  * Add an `Autowire` attribute to tell a parameter how to be autowired
  * Allow using expressions as service factories
+ * Add argument type `closure` to help passing closures to services
  * Deprecate `ReferenceSetArgumentTrait`
 
 6.0